Поменял тип поля quantity на BigDecimal

+++ backend/src/main/java/com/github/giga_chill/gigachill/data/transfer/object/ShoppingItemDTO.java
@@ -2,9 +2,11 @@ package com.github.giga_chill.gigachill.data.transfer.object;
 
 import jakarta.annotation.Nullable;
 
+import java.math.BigDecimal;
+
 public record ShoppingItemDTO(String shoppingItemId,
                               @Nullable String title,
-                              @Nullable Integer quantity,
+                              @Nullable BigDecimal quantity,
                               @Nullable String unit,
                               @Nullable Boolean isPurchased) {
 }
